+++ src/features/StorePage/components/StoreContentHeader.js
@@ -22,5 +22,3 @@ function StoreContentHeader() {
     </div>
   );
 }
-
-export default StoreContentHeader;
