+++ src/features/StorePage/components/StoreFeaturedContent.css
