+++ src/features/StorePage/index.js
@@ -1,0 +1,3 @@
+export { default as StoreContentHeader } from './components/StoreContentHeader.js';
+export { default as StoreFeaturedContent } from './components/StoreFeaturedContent.js';
+
